Added database migration using FlyWay, changed the id type to the entity Article from Integer to Long.

+++ src/main/java/com/krupoderov/spring/articles/controller/ArticlesController.java
@@ -57,7 +57,7 @@ public class ArticlesController {
      * @return переадрессация
      */
     @RequestMapping(value = "/editor/delete/{article_id}")
-    public String deleteArticle(@PathVariable("article_id") Integer articleId) {
+    public String deleteArticle(@PathVariable("article_id") Long articleId) {
         service.delete(articleId);
         return "redirect:/";
     }
